Separate the signed response from the unsigned ones.
Trying to keep this within the same class is like calling out for
problems with response trickery and a messy code. The Signed response
code is really just the same as for the non-signed, but looking for data
in the better places. Also implementing the ->getData() call getting the
signed data from the response. This should be used always really here.