Do not always vacuum the data this will mess up signature verification in responses.
Always vacuuming data will remove all null values in the incoming data
request, this will in turn make the verification of the incoming
signature fail (as the keys will need to be in the serialized data). So
stop vacuuming all data, make the Data class only create an empty object
and let the children do vacuuming as needed.